some guards and remove extra files in travis

+++ .travis/deploy_script.sh
@@ -7,6 +7,12 @@ else
 fi
 
 echo $PKG_NAME
-#rm -rf .git deploy_rsa.enc .idea phpvms.iml
-tar -czf $PKG_NAME.tar.gz -C $TRAVIS_BUILD_DIR .
-rsync -r --delete-after --quiet $PKG_NAME.tar.gz downloads@phpvms.net:/var/www/downloads/
+
+if [ "$TRAVIS" = "true" ]; then
+    echo "Runnign on travis"
+    rm -rf .git deploy_rsa.enc .idea phpvms.iml .travis .dpl
+    find . -type d -name ".git" -delete
+
+    tar -czf $PKG_NAME.tar.gz -C $TRAVIS_BUILD_DIR .
+    rsync -r --delete-after --quiet $PKG_NAME.tar.gz downloads@phpvms.net:/var/www/downloads/
+fi
